src/test/java/com/helloworld/HelloWorldServletTest.java
package com.helloworld;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.junit.jupiter.MockitoExtension;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;
import java.io.PrintWriter;
import java.io.StringWriter;

import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class HelloWorldServletTest {


    @Test
    void shouldReturn_HelloWorld_Response() throws IOException {

        HttpServletRequest request = mock(HttpServletRequest.class);
        HttpServletResponse response = mock(HttpServletResponse.class);

        StringWriter stringWriter = new StringWriter();
        PrintWriter writer = new PrintWriter(stringWriter);
        when(response.getWriter()).thenReturn(writer);

        new HelloWorldServlet().doGet(request, response);

        assertTrue(stringWriter.toString().equals("<style> h1 {background-color: #2D2D2D; color: #C26356; font-size: 50px;" +
                "font-family: Menlo, Monaco, fixed-width;} </style><h1> Hello World</h1>"));
    }


}